+++ src/packages/result-list/ResultList.tsx
@@ -3,8 +3,8 @@
 import { Box } from "@open-pioneer/chakra-integration";
 import { CommonComponentProps, useCommonComponentProps } from "@open-pioneer/react-utils";
 import { Dispatch, FC, SetStateAction, useEffect, useMemo, useRef, useState } from "react";
-import { DataTable } from "./DataTable";
 import { ResultListInput, ResultListSelectionChangedEvent } from "./api";
+import { DataTable } from "./DataTable/DataTable";
 import { createColumns } from "./createColumns";
 import { useIntl } from "open-pioneer:react-hooks";
 import { BaseFeature } from "@open-pioneer/map/api/BaseFeature";
@@ -27,27 +27,18 @@ export const ResultList: FC<ResultListProps> = (props) => {
     const { resultListInput, getSelectedFeature, onSelectionChanged } = props;
     const data = resultListInput.data;
     const metadata = resultListInput.metadata;
-
     if (metadata.length === 0) {
-        throw Error(intl.formatMessage({ id: "illegalArgumentException" }));
+        throw Error("no metadata given! result list cannot be displayed.");
     }
 
     const dataTableRef = useRef<HTMLDivElement>(null);
-    const [tableWidth, setTableWidth] = useState(0);
-
-    useEffect(() => {
-        if (!dataTableRef.current) return;
-        const resizeObserver = new ResizeObserver((event) => {
-            // Depending on the layout, you may need to swap inlineSize with blockSize
-            // https://developer.mozilla.org/en-US/docs/Web/API/ResizeObserverEntry/contentBoxSize
-            if (event[0] === undefined || event[0].contentBoxSize[0] === undefined) return;
-            setTableWidth(event[0].contentBoxSize[0].inlineSize);
-        });
-        resizeObserver.observe(dataTableRef.current);
-    }, [dataTableRef.current]);
+    const tableWidth = useTableWidth(dataTableRef);
 
     const { containerProps } = useCommonComponentProps("result-list", props);
-    const columns = useMemo(() => createColumns(metadata, tableWidth), [metadata, tableWidth]);
+    const columns = useMemo(
+        () => createColumns(metadata, intl, tableWidth),
+        [metadata, intl, tableWidth]
+    );
 
     return (
         <Box {...containerProps} height="100%" overflowY="auto" ref={dataTableRef}>
@@ -60,3 +51,20 @@ export const ResultList: FC<ResultListProps> = (props) => {
         </Box>
     );
 };
+
+function useTableWidth(tableRef: React.RefObject<HTMLDivElement> | null) {
+    const [tableWidth, setTableWidth] = useState(0);
+
+    useEffect(() => {
+        if (!tableRef?.current) return;
+        const resizeObserver = new ResizeObserver((event) => {
+            // Depending on the layout, you may need to swap inlineSize with blockSize
+            // https://developer.mozilla.org/en-US/docs/Web/API/ResizeObserverEntry/contentBoxSize
+            if (event[0] === undefined || event[0].contentBoxSize[0] === undefined) return;
+            setTableWidth(event[0].contentBoxSize[0].inlineSize);
+        });
+        resizeObserver.observe(tableRef.current);
+    }, [tableRef?.current]);
+
+    return tableWidth;
+}