+++ src/packages/result-list/createColumns.test.ts
@@ -3,17 +3,22 @@
 
 import { afterEach, expect, vi, it, describe } from "vitest";
 import { ResultColumn } from "./api";
-import { createColumns } from "./createColumns";
+import { createColumns, SELECT_COLUMN_SIZE } from "./createColumns";
+import { dummyMetaDataMissingWidth } from "./testSources";
+import { createIntl } from "@open-pioneer/test-utils/vanilla";
 
 afterEach(() => {
     vi.restoreAllMocks();
 });
 
 describe("create columns for result-list", () => {
+    const intl = createIntl();
+
     it("expect createColumn to create Columns correctly", async () => {
         const metaData = createMetaData();
+
         // Slice away the selection checkbox column
-        const columns = createColumns(metaData).slice(1);
+        const columns = createColumns(metaData, intl).slice(1);
         expect(columns.length).toEqual(metaData.length);
         columns.forEach((column, index) => {
             //Column-ID is correct
@@ -27,7 +32,17 @@ describe("create columns for result-list", () => {
             // no column-size testing because of recalculation based on column definition
         });
     });
-    // TODO: Add tests for width calculation?
+    it("expect createColumn to distribute remaining width on columns with undefined width", async () => {
+        const metaData = dummyMetaDataMissingWidth;
+        const fullWidth = 1000;
+        // Slice away the selection checkbox column
+        const columns = createColumns(metaData, intl, fullWidth).slice(1);
+        const expectedWidth = (fullWidth - SELECT_COLUMN_SIZE - 300) / 2;
+        expect(columns[0]?.size).toEqual(metaData[0]!.width);
+        expect(columns[1]?.size).toEqual(expectedWidth);
+        expect(columns[2]?.size).toEqual(expectedWidth);
+        expect(columns[3]?.size).toEqual(metaData[3]!.width);
+    });
 });
 
 function createMetaData() {