+++ src/samples/map-sample/ol-app/ui/ResultList.tsx
@@ -6,7 +6,7 @@ import { ResultList } from "@open-pioneer/result-list";
 import { AppModel } from "../AppModel";
 import { useSnapshot } from "valtio";
 import { useService } from "open-pioneer:react-hooks";
-import { useState } from "react";
+import { useCallback, useState } from "react";
 import { BaseFeature } from "@open-pioneer/map/api/BaseFeature";
 import { ResultListSelectionChangedEvent } from "@open-pioneer/result-list/api";
 
@@ -14,12 +14,11 @@ export function ResultListComponent() {
     const appModel = useService<unknown>("ol-app.AppModel") as AppModel;
     const input = useSnapshot(appModel.state).currentResultListInput;
     const [selectedFeatures, setSelectedFeatures] = useState<null | BaseFeature[]>(null);
+    const onSelectionChanged = useCallback((_event: ResultListSelectionChangedEvent) => {
+        console.log("changed");
+    }, []);
     if (!input) {
         return undefined;
-    }
-
-    function onSelectionChanged(_event: ResultListSelectionChangedEvent) {
-        console.log("changed");
     }
 
     return (