+++ src/packages/search/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @open-pioneer/search
 
+## 0.3.1
+
+### Patch Changes
+
+-   611ddb9: Export interface `BaseFeature` from Map API and use it correctly in base packages `selection` and `search`.
+-   Updated dependencies [611ddb9]
+    -   @open-pioneer/map@0.3.1
+
+## 0.3.0
+
+### Minor Changes
+
+-   ee7c2d4: Update runtime version.
+-   0456500: Add interface `BaseFeature` to Map API.
+
+### Patch Changes
+
+-   0883bbd: Add property `menuPosition` to React Select.
+-   Updated dependencies [ee7c2d4]
+-   Updated dependencies [a582e5e]
+-   Updated dependencies [0456500]
+-   Updated dependencies [762e7b9]
+    -   @open-pioneer/map@0.3.0
+    -   @open-pioneer/react-utils@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes