+++ src/samples/test-toc/toc-app/CHANGELOG.md
@@ -1,5 +1,26 @@
 # toc-app
 
+## 0.0.5
+
+### Patch Changes
+
+-   Updated dependencies [611ddb9]
+    -   @open-pioneer/map@0.3.1
+    -   @open-pioneer/toc@0.3.1
+
+## 0.0.4
+
+### Patch Changes
+
+-   Updated dependencies [0883bbd]
+-   Updated dependencies [ee7c2d4]
+-   Updated dependencies [a582e5e]
+-   Updated dependencies [0456500]
+-   Updated dependencies [762e7b9]
+    -   @open-pioneer/toc@0.3.0
+    -   @open-pioneer/map@0.3.0
+    -   @open-pioneer/react-utils@0.2.1
+
 ## 0.0.3
 
 ### Patch Changes